refactor: replace all accessMode with utility accessMode
