+++ src/verify-request/login-again-if-different-shop.ts
@@ -3,7 +3,8 @@ import {Context} from 'koa';
 import Shopify from '@shopify/shopify-api';
 import { Session } from '@shopify/shopify-api/dist/auth/session';
 
-import {AccessMode, NextFunction} from '../types';
+import {NextFunction} from '../types';
+import {AccessMode, isAccessModeOnline} from '../utilities'
 
 import {Routes} from './types';
 import {clearSession, redirectToAuth} from './utilities';
@@ -16,7 +17,7 @@ export function loginAgainIfDifferentShop(routes: Routes, accessMode: AccessMode
   ) {
     const {query} = ctx;
     let session: Session | undefined;
-    session = await Shopify.Utils.loadCurrentSession(ctx.req, ctx.res, accessMode === 'online');
+    session = await Shopify.Utils.loadCurrentSession(ctx.req, ctx.res, isAccessModeOnline(accessMode));
 
     if (session && query.shop && session.shop !== query.shop) {
       await clearSession(ctx, accessMode);
