+++ src/verify-request/utilities.ts
@@ -3,7 +3,7 @@ import {Context} from 'koa';
 import Shopify from '@shopify/shopify-api';
 
 import {Routes} from './types';
-import {AccessMode} from '../types';
+import {AccessMode, isAccessModeOnline} from '../utilities';
 import {DEFAULT_ACCESS_MODE} from '../auth';
 
 export function redirectToAuth(
@@ -22,7 +22,7 @@ export function redirectToAuth(
 
 export async function clearSession(ctx: Context, accessMode: AccessMode = DEFAULT_ACCESS_MODE) {
   try {
-    await Shopify.Utils.deleteCurrentSession(ctx.req, ctx.res, accessMode === 'online');
+    await Shopify.Utils.deleteCurrentSession(ctx.req, ctx.res, isAccessModeOnline(accessMode));
   }
   catch (error) {
     if (error instanceof Shopify.Errors.SessionNotFound) {
